Change node-mongo api helpers location and imports
