+++ esm/node-mongo-helpers.js
@@ -1,17 +1,31 @@
-const { spawn } = require('child_process');
-const { success, error, warning } = require('./consolemsg');
-const package_json = require('../../package.json');
+import { spawn } from 'child_process';
+import package_json from './package.json';
+import chalk from 'chalk';
 
 /* eslint-disable no-console */
 
-const watchEslint = () => {
+// console
+export const success = (message) => {
+  console.log( chalk.greenBright(message) );
+}
+
+export const warning = (message) => {
+  console.log( chalk.yellowBright(message) );
+}
+
+export const error = (message) => {
+  console.log( chalk.redBright(message) );
+}
+
+// DB connect
+export const watchEslint = () => {
   const npm = process.platform === 'win32' ? 'npm.cmd' : 'npm';
   spawn(npm, ['run', 'lint:watch'], { cwd: './', stdio: 'inherit' });
 }
 
-const server = (serverPort) => {
+export const server = (serverPort) => {
   try {
-    success(`\nnode-mongo (Commonjs Module) API boilerplate template v${package_json.version}`);
+    success(`\nnode-mongo (ES Module) API boilerplate template v${package_json.version}`);
     success(`\nServer running at ${serverPort}`);
   } catch (err) {
     error(`${{ err }}`);
@@ -23,25 +37,17 @@ const eslintAndServer = (serverPort) => {
   server(serverPort);
 }
 
-const afterAtlasDBconnectSuccessful = (serverPort) => {
+export const afterAtlasDBconnectSuccessful = (serverPort) => {
   success('\nConnected to mongoDB ATLAS');
   eslintAndServer(serverPort);
 }
 
-const afterLocalDBconnectSuccessful = (serverPort) => {
+export const afterLocalDBconnectSuccessful = (serverPort) => {
   success('\nConnected to LOCAL mongoDB');
   eslintAndServer(serverPort);
 }
 
-const connectToDBunsuccessful = (err) => {
+export const connectToDBunsuccessful = (err) => {
   error(`\nError in DB connection: ${err.message}\n`);
   warning('Refer to the node-mongo documentation: https://code-collabo.gitbook.io/node-mongo-v2.0.0\n\nGet further help from Code Collabo Community\'s Node mongo channel: https://matrix.to/#/#collabo-node-mongo:gitter.im\n');
 }
-
-module.exports = {
-  watchEslint,
-  server,
-  afterAtlasDBconnectSuccessful,
-  afterLocalDBconnectSuccessful,
-  connectToDBunsuccessful,
-};
