+++ esm/src/api/controllers/demo.controller.js
@@ -6,7 +6,7 @@ import {
   updateOneDemoItemPropertyValueService,
   updateDemoItemPropertyValuesService,
 } from '../services/demo.service';
-import { success, error } from '../../lib/consolemsg';
+import { success, error } from '../../../node-mongo-helpers';
 
 const routeName = 'demo';
 const item = `${routeName}-item`;
