+++ ts/src/api/controllers/demo.controller.ts
@@ -7,7 +7,7 @@ import {
   updateOneDemoItemPropertyValueService,
   updateDemoItemPropertyValuesService,
 } from '../services/demo.service';
-import { success, error } from '../../lib/consolemsg';
+import { success, error } from '../../../node-mongo-helpers';
 
 const routeName = 'demo';
 const item = `${routeName}-item`;
