--- ts/src/lib/consolemsg.ts
@@ -1,15 +1,0 @@
-import chalk from 'chalk';
-
-/* eslint-disable no-console */
-
-export const success = (message: string) => {
-  console.log( chalk.greenBright(message) );
-}
-
-export const warning = (message: string) => {
-  console.log( chalk.yellowBright(message) );
-}
-
-export const error = (message: string) => {
-  console.log( chalk.redBright(message) );
-}
